+++ src/routes/habitat/components/CameraControlModal/index.js
@@ -3,11 +3,8 @@ import { Layer, Box } from 'grommet';
 
 import Header from 'Components/modals/Header';
 import { showSwitchCameraModal } from './actions';
-import { POWER, SWITCH } from './types';
+import { POWER } from './types';
 import Power from './Power';
-import SwitchStream from './SwitchStream';
-
-import style from './style.scss';
 
 const CameraControlModal = ({ show, type, showSwitchCameraModalAction }) => {
   if (!show) {
@@ -17,16 +14,14 @@ const CameraControlModal = ({ show, type, showSwitchCameraModalAction }) => {
   const onClose = () => showSwitchCameraModalAction(false);
 
   return (
-    <Layer position="center" onClickOutside={onClose} onEsc={onClose} className={style.cameraControlModal}>
+    <Layer position="center" onClickOutside={onClose} onEsc={onClose}>
       <Box height={{ max: 'calc((100 * var(--vh)) - 30px)'}}>
         <Header onClose={onClose}>
           {type === POWER && 'Power'}
-          {type === SWITCH && 'Switch Stream'}
         </Header>
 
         <Box height="100%">
           {type === POWER && <Power />}
-          {type === SWITCH && <SwitchStream />}
         </Box>
       </Box>
     </Layer>