+++ src/routes/habitat/components/CameraControlModal/types.js
@@ -1,3 +1,2 @@
 export const SHOW_CAMERA_CONTROL_MODAL = 'SHOW_CAMERA_CONTROL_MODAL';
 export const POWER = 'POWER';
-export const SWITCH = 'SWITCH';